+++ plugins/items/barrierDamage.lua
@@ -1,8 +1,8 @@
--- Overloaded Capacitor
+-- Barrier Damage
 
 local sprite = Resources.sprite_load(PATH.."assets/sprites/overloadedCapacitor.png", 1, false, false, 16, 16)
 
-local item = Item.create("aphelion", "overloadedCapacitor")
+local item = Item.create("aphelion", "barrierDamage")
 Item.set_sprite(item, sprite)
 Item.set_tier(item, Item.TIER.uncommon)
 Item.set_loot_tags(item, Item.LOOT_TAG.category_damage)
@@ -12,9 +12,9 @@ Item.add_callback(item, "onStep", function(actor, stack)
         Actor.set_barrier(actor, math.max(actor.barrier, actor.maxbarrier * 0.08))
     end
 
-    if actor.barrier > 0 then
-        Buff.apply(actor, Buff.find("aphelion-overloadedCapacitor"), 2)
-    end
+    -- if actor.barrier > 0 then
+    --     Buff.apply(actor, Buff.find("aphelion-overloadedCapacitor"), 2)
+    -- end
 end)
 
 Item.add_callback(item, "onAttack", function(actor, damager, stack)
@@ -28,7 +28,7 @@ end)
 
 -- Buff
 
-local sprite = Resources.sprite_load(PATH.."assets/sprites/buffOverloadedCapacitor.png", 1, false, false, 7, 7)
+-- local sprite = Resources.sprite_load(PATH.."assets/sprites/buffOverloadedCapacitor.png", 1, false, false, 7, 7)
 
-local buff = Buff.create("aphelion", "overloadedCapacitor")
-Buff.set_property(buff, Buff.PROPERTY.icon_sprite, sprite)
+-- local buff = Buff.create("aphelion", "overloadedCapacitor")
+-- Buff.set_property(buff, Buff.PROPERTY.icon_sprite, sprite)