+++ CHANGELOG.md
@@ -8,10 +8,11 @@
 * Six Shooter now works with Sniper's Snipe properly (fixed in RoRR Modding Toolkit).
 * Ballistic Vest : Is now achievement locked
 * Overloaded Capacitor
+    * Renamed to Barrier Damage (temp name)
     * First stack damage bonus: 20% -> 15%
     * Barrier out-of-combat delay time: 3s -> 5s
     * Barrier out-of-combat gain: 6% -> 8%
 * Stiletto : Is now achievement locked
 * Whimsical Star : Is now achievement locked
 * Added Relic Guard
-* Added Super Shield (temp name - also functionality changed from doc)
+* Added Overloaded Capacitor (new version)